v0.1.4

+++ pages/images/db_models.py
@@ -6,7 +6,7 @@ class ImagesLibrary(db.Model):
   id = db.Column(db.Integer, unique=True, primary_key=True)
   hash = db.Column(db.String, nullable=True, unique=True) #, index=True
   file_path = db.Column(db.String, nullable=True)
-  user_rating = db.Column(db.Integer, nullable=True)
+  user_rating = db.Column(db.Float, nullable=True)
   user_rating_date = db.Column(db.DateTime, nullable=True)
   model_rating = db.Column(db.Float, nullable=True)
   model_hash = db.Column(db.String, nullable=True)
